!: proper handling of connected instruments on re-connect
http://r2d2.hq.sarad.de:3000/SARAD/RegServer/issues/38#issuecomment-378
Freeing all instruments on re-connect is somewhat dirty, given the fact that the
IS can have multiple frontends and the instrument can be reserved over every of
those frontends. This patch frees all instrument on every (re-)connect to MQTT
broker. This is acceptable if we do not encourage using mDNS and MQTT frontend
in parallel.